requirements.txt
python-telegram-bot>=20.0
pyyaml>=6.0
sqlalchemy>=2.0
aiosqlite>=0.19
pydantic>=2.0
python-dotenv>=1.0
